Make needs_dynamic_casting multiple-complex-type aware.
We don't actually cast between these complex representations, but the prior implementation would indicate that we needed to dynamic_cast,
because we didn't have mappings for std::complex or thrust::complex.

This PR makes it so they all map to the same dtype.

Note that this has no functional change as all the use sites have already been changed to take this into account.

ghstack-source-id: fcc12c06e18362f9f2615091eff0a74184593936
Pull Request resolved: https://github.com/pytorch/pytorch/pull/39255